fix batchsize bug and muiti-gpus training

+++ look2hear/models/apollo.py
@@ -292,7 +292,7 @@ class Apollo(BaseModel):
         for i in range(self.nband):
             this_RI = self.output[i](feature[:,i]).view(B*nch, 2, self.band_width[i], -1).float()
             est_spec.append(torch.complex(this_RI[:,0], this_RI[:,1]))
-        est_spec = torch.cat(est_spec, 1)
+        est_spec = torch.cat(est_spec, 1).to(dtype=torch.complex64)
         output = torch.istft(est_spec, n_fft=self.win, hop_length=self.stride, 
                              window=torch.hann_window(self.win).to(input.device), length=nsample).view(B, nch, -1)
